refactor(types): consolidate bitbucket types into types.ts
